+++ R/use_bst_file.R
@@ -7,10 +7,8 @@
 #'
 #' @inheritParams starter::use_project_file
 #' @inheritParams create_bst_project
-#' @name use_bst_file
-#' @rdname use_bst_file
+#' @name use_file
 #' @seealso [`create_bst_project()`]
-#' @export
 #' @examples
 #' \donttest{\dontrun{
 #' # create gitignore file
@@ -21,11 +19,14 @@
 #' use_project_file("readme")
 #' use_project_readme()
 #' }}
+NULL
 
+#' @rdname use_file
+#' @export
 use_bst_file <- function(name = NULL,
-                             filename = NULL,
-                             template = NULL,
-                             open = interactive()) {
+                         filename = NULL,
+                         template = NULL,
+                         open = interactive()) {
   # if template is NULL, use default template ----------------------------------
   template <-
     template %||%
@@ -36,17 +37,62 @@ use_bst_file <- function(name = NULL,
                             template = template, open = open)
 }
 
-#' @rdname use_bst_file
+#' @rdname use_file
 #' @export
 use_bst_gitignore <- function(filename = NULL, template = NULL) {
   use_bst_file(name = "gitignore", filename = filename,
-                   template = template)
+               template = template)
 }
 
-#' @rdname use_bst_file
+#' @rdname use_file
 #' @export
 use_bst_readme <- function(filename = NULL, template = NULL) {
   use_bst_file(name = "readme", filename = filename,
-                   template = template)
+               template = template)
 }
+
+#' @rdname use_file
+#' @export
+use_hot_file <- function(name = NULL, filename = NULL,
+                         template = bstfun::project_templates[["hot"]],
+                         open = interactive()) {
+  starter::use_project_file(name = name, filename = filename,
+                            template = template, open = open)
+}
+
+#' @rdname use_file
+#' @export
+use_hot_gitignore <- function(filename = NULL, template = NULL) {
+  use_hot_file(name = "gitignore", filename = filename,
+               template = template)
+}
+
+#' @rdname use_file
+#' @export
+use_hot_readme <- function(filename = NULL, template = NULL) {
+  use_hot_file(name = "readme", filename = filename,
+               template = template)
+}
+
+#' @rdname use_file
+#' @export
+use_hot_setup <- function(filename = NULL, template = NULL) {
+  use_hot_file(name = "setup", filename = filename,
+               template = template)
+}
+
+#' @rdname use_file
+#' @export
+use_hot_analysis <- function(filename = NULL, template = NULL) {
+  use_hot_file(name = "analysis", filename = filename,
+               template = template)
+}
+
+#' @rdname use_file
+#' @export
+use_hot_report <- function(filename = NULL, template = NULL) {
+  use_hot_file(name = "report", filename = filename,
+               template = template)
+}
+