+++ NEWS.md
@@ -1,3 +1,7 @@
+# bstfun (development version)
+
+* Removed use of deprecated function in `tbl_likert()`.
+
 # bstfun 0.2.8
 
 * Updated how references are added in `cite_r()`.