+++ NEWS.md
@@ -1,5 +1,11 @@
 # bstfun (development version)
 
+* Added new function `clean_mrn()` to clean and check consistency of MRN columns.
+
+* Added new function `count_map()` to aid in checking variable derivations.
+
+* Added new function `assign_timepoint()` to aid in selecting observations at various follow-up times, e.g. selecting the 6 and 12 month lab result from a long data set of all labs.
+
 * Removed use of deprecated function in `tbl_likert()`.
 
 * Added new function `followup_time()` that reports the amount of follow-up among censored patients.