+++ NEWS.md
@@ -1,4 +1,8 @@
-# bstfun (development version)
+# bstfun 0.3.0
+
+* Function `add_cuminc_risktables()` now allows for 3 groups. Makes colors consistent between labels and estimates.  Removed "Number at risk" label and places a label ", n:" next to each strata name. 
+
+* Migrated the H.O.T. project templates from {hotfun} to {bstfun}.
 
 * Added new function `clean_mrn()` to clean and check consistency of MRN columns.